mockito-1 - Renaming packages
