+++ junit-examples/src/test/java/com/hargun/unittesting/junit/ExceptionExampleTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.hargun.junit.lecture1;
+package com.hargun.unittesting.junit;
 
 import java.util.Arrays;
