+++ junit-examples/src/test/java/com/hargun/unittesting/junit/PerformanceExampleTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.hargun.junit.lecture1;
+package com.hargun.unittesting.junit;
 
 import java.util.Arrays;
